Update IBM.MQ.xml
Added installPath from server class property as parameter in channel/listener/queue discovery to run commands even if installPath is not in env.variables.
Therefore updated the commands (including quotes because of space in "Program Files"). Dot-Executing binary command (dspmq) instead of Invoke-Express (not working here with space in path).
Added Set-WinUILanguageOverride to allow the script to run with non-"en-US" locale (in de-* for instance the channel status is "Aktiv" instead of "RUNNING"). Unfortunately, for executables it is not possible to set the culture for the thread/session and therefore it may be set temporary using Set-WinUILanguageOverride.
These changes are tested and verified.
Unsealed XML only, please seal MP with your own key.